requirements.txt
asgiref==3.8.1
# backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==42.0.8
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
pandas==2.2.2
neo4j